Don't link the shell to the shim in detached mode
In case our runtime is used in standalone mode with the --detach
flag enabled, we should not get any output on the current shell.
We should still be able to get some outputs if a console has been
provided.

This patch relies on the latest virtcontainers vendoring to be
able to indicate if we are in case of a detached command or not.

Fixes #326

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>